SeguimientoProyectos/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
celery>=5.3,<6.0
reportlab>=4.0,<5.0